Use auto diff to get at jacobian of deflection angles
These changes add the ability to use JAX to find the jacobian of deflection angles.  To do this it first needs a version of the deflection angle function that takes in two scalars and returns a 2 element vector.  From this `jax.jacfwd` is used to get the 2x2 jacobian for a single (y,x) value.  This function is vectorized to take in a set of (y,x) values (of arbitrary shape) and return a (*shape, 2, 2) array of jacobian values.

The final step is to pass the value from a `grid` object into this function.